refactor(toolkit): split identity into users + groups resources
Carve the monolithic `experimental.clients.identity` package into two proper
resource packages and leave `clients.identity` as a thin facade, as
scaffolded in the reorg proposal.

- `experimental/resources/users/` now owns
  - schemas.py: `UserInfo`, `UserWithConfiguration`, `UserGroupMembership`
  - functions.py: `list_users`, `get_user_by_id`, `find_user`,
    `get_user_groups`, `update_user_configuration` (+ `_async` siblings)
    together with their private helpers and pagination constants
  - service.py: `Users` class, including the relational `groups_of` /
    `is_member` helpers (which stay on `Users` because the SDK endpoint
    is rooted on the user) and the `_resolve_user_identifier` helper
- `experimental/resources/groups/` now owns
  - schemas.py: `GroupInfo`, `GroupWithConfiguration`, `GroupMember`,
    `GroupMembership`, `GroupDeleted`
  - functions.py: `list_groups`, `create_group`, `delete_group`,
    `rename_group`, `update_group_configuration`, `add_group_members`,
    `remove_group_members` (+ `_async` siblings) and their helpers
  - service.py: `Groups` class
- `experimental/clients/identity/` is reduced to the `Identity` facade:
  - service.py now only defines `Identity`, which composes `Users` and
    `Groups` from the two resource packages and keeps `from_context` /
    `from_settings` constructors
  - __init__.py re-exports `Identity` plus `Users`/`Groups` and the schemas
    from the resource packages so existing `from
    unique_toolkit.experimental.clients.identity import ...` call sites
    keep resolving without pulling in a second source of truth
  - schemas.py and functions.py are removed (their content now lives in
    the resource packages)

Verified that all three new top-level packages import cleanly and that
`Users is ...clients.identity.Users` / `Groups is ...clients.identity.Groups`
resolve to the resource classes (no accidental duplication). `ruff check`
passes for all three packages.

External callers (tests under `tests/identity/` and docs under
`docs/.python_files/` + `docs/examples_from_docs/`) still import from the
pre-move path `unique_toolkit.experimental.identity` and remain broken in
the same way they were before this commit; updating them is out of scope
and tracked separately.

Made-with: Cursor
