+++ unique_toolkit/unique_toolkit/experimental/resources/groups/__init__.py
@@ -1,16 +1,70 @@
 """Typed wrapper over the ``unique_sdk`` group resource.
 
-Carved out of the current monolithic ``identity/`` package. Only the
-group-centric pieces live here:
+Carved out of the original monolithic ``experimental.identity`` package.
+Only the group-centric pieces live here:
 
-* ``Groups`` service (from ``identity/service.py``)
-* functions: ``list_groups``, ``create_group``, ``delete_group``,
-  ``rename_group``, ``update_group_configuration``, ``add_group_members``,
-  ``remove_group_members`` (+ their ``_async`` siblings, from
-  ``identity/functions.py``)
-* schemas: ``GroupInfo``, ``GroupWithConfiguration``, ``GroupMember``,
-  ``GroupMembership``, ``GroupDeleted`` (from ``identity/schemas.py``)
+* :class:`Groups` service (from :mod:`.service`)
+* functions: :func:`list_groups`, :func:`create_group`, :func:`delete_group`,
+  :func:`rename_group`, :func:`update_group_configuration`,
+  :func:`add_group_members`, :func:`remove_group_members` (plus their
+  ``_async`` siblings, from :mod:`.functions`)
+* schemas: :class:`GroupInfo`, :class:`GroupWithConfiguration`,
+  :class:`GroupMember`, :class:`GroupMembership`, :class:`GroupDeleted`
+  (from :mod:`.schemas`)
 
-The :class:`Identity` facade that composes ``Users`` + ``Groups`` lives in
-:mod:`..clients.identity`.
+The :class:`Identity` facade that composes
+:class:`~unique_toolkit.experimental.resources.users.Users` + :class:`Groups`
+lives in :mod:`unique_toolkit.experimental.clients.identity`.
 """
+
+from unique_toolkit.experimental.resources.groups.functions import (
+    DEFAULT_LIST_SKIP,
+    DEFAULT_LIST_TAKE,
+    add_group_members,
+    add_group_members_async,
+    create_group,
+    create_group_async,
+    delete_group,
+    delete_group_async,
+    list_groups,
+    list_groups_async,
+    remove_group_members,
+    remove_group_members_async,
+    rename_group,
+    rename_group_async,
+    update_group_configuration,
+    update_group_configuration_async,
+)
+from unique_toolkit.experimental.resources.groups.schemas import (
+    GroupDeleted,
+    GroupInfo,
+    GroupMember,
+    GroupMembership,
+    GroupWithConfiguration,
+)
+from unique_toolkit.experimental.resources.groups.service import Groups
+
+__all__ = [
+    "DEFAULT_LIST_SKIP",
+    "DEFAULT_LIST_TAKE",
+    "GroupDeleted",
+    "GroupInfo",
+    "GroupMember",
+    "GroupMembership",
+    "GroupWithConfiguration",
+    "Groups",
+    "add_group_members",
+    "add_group_members_async",
+    "create_group",
+    "create_group_async",
+    "delete_group",
+    "delete_group_async",
+    "list_groups",
+    "list_groups_async",
+    "remove_group_members",
+    "remove_group_members_async",
+    "rename_group",
+    "rename_group_async",
+    "update_group_configuration",
+    "update_group_configuration_async",
+]
